# backend/src/assignments/dto/update-assignment.dto.ts
import { ApiPropertyOptional } from '@nestjs/swagger';
import { IsDateString, IsOptional, IsString } from 'class-validator';

export class UpdateAssignmentDto {
  @ApiPropertyOptional({ example: 'Homework 2' })
  @IsString()
  @IsOptional()
  title?: string;

  @ApiPropertyOptional({ example: '2026-01-15' })
  @IsDateString()
  @IsOptional()
  dueDate?: string;
}