# openapi.json
{"openapi":"3.0.0","paths":{"/auth/register":{"post":{"operationId":"AuthController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDto"}}}},"responses":{"201":{"description":"Returns a JWT access token"},"409":{"description":"Email already registered"}},"summary":"Register a new user account","tags":["auth"]}},"/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"200":{"description":"Returns a JWT access token"},"401":{"description":"Invalid credentials"}},"summary":"Login and receive a JWT access token","tags":["auth"]}},"/students":{"get":{"operationId":"StudentsController_findAll","parameters":[],"responses":{"200":{"description":"Array of student objects"}},"security":[{"bearer":[]}],"summary":"List all students with profiles and enrolled courses","tags":["students"]},"post":{"operationId":"StudentsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateStudentDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Student"}}}},"409":{"description":"Email already in use"}},"security":[{"bearer":[]}],"summary":"[Admin] Create a new student","tags":["students"]}},"/students/enrollments":{"get":{"operationId":"StudentsController_findAllEnrollments","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Student"}}}}}},"security":[{"bearer":[]}],"summary":"List all students with their course enrollments","tags":["students"]}},"/students/{id}":{"get":{"operationId":"StudentsController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Student"}}}},"404":{"description":"Student not found"}},"security":[{"bearer":[]}],"summary":"Get a single student with profile and courses","tags":["students"]},"patch":{"operationId":"StudentsController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateStudentDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Student"}}}}},"security":[{"bearer":[]}],"summary":"[Admin] Update student details","tags":["students"]},"delete":{"operationId":"StudentsController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"204":{"description":""}},"security":[{"bearer":[]}],"summary":"[Admin] Soft-delete a student (recoverable)","tags":["students"]}},"/students/{studentId}/enroll/{courseId}":{"post":{"operationId":"StudentsController_enroll","parameters":[{"name":"studentId","required":true,"in":"path","schema":{"type":"number"}},{"name":"courseId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Student"}}}},"409":{"description":"Student already enrolled in this course"}},"security":[{"bearer":[]}],"summary":"[Admin] Enroll a student in a course","tags":["students"]}},"/students/{studentId}/unenroll/{courseId}":{"delete":{"operationId":"StudentsController_unenroll","parameters":[{"name":"studentId","required":true,"in":"path","schema":{"type":"number"}},{"name":"courseId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Student"}}}}},"security":[{"bearer":[]}],"summary":"[Admin] Remove a student from a course","tags":["students"]}},"/profiles":{"get":{"operationId":"ProfilesController_findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Profile"}}}}}},"security":[{"bearer":[]}],"summary":"List all profiles","tags":["profiles"]},"post":{"operationId":"ProfilesController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateProfileDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Profile"}}}},"409":{"description":"Student already has a profile"}},"security":[{"bearer":[]}],"summary":"[Admin] Create a profile for a student","tags":["profiles"]}},"/profiles/{id}":{"get":{"operationId":"ProfilesController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Profile"}}}},"404":{"description":"Profile not found"}},"security":[{"bearer":[]}],"summary":"Get a single profile","tags":["profiles"]},"patch":{"operationId":"ProfilesController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProfileDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Profile"}}}}},"security":[{"bearer":[]}],"summary":"[Admin] Update a profile","tags":["profiles"]},"delete":{"operationId":"ProfilesController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"204":{"description":""}},"security":[{"bearer":[]}],"summary":"[Admin] Delete a profile","tags":["profiles"]}},"/courses":{"get":{"operationId":"CoursesController_findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Course"}}}}}},"security":[{"bearer":[]}],"summary":"List all courses with assignments and enrolled students","tags":["courses"]},"post":{"operationId":"CoursesController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCourseDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Course"}}}},"409":{"description":"Course code already exists"}},"security":[{"bearer":[]}],"summary":"[Admin] Create a new course","tags":["courses"]}},"/courses/{id}":{"get":{"operationId":"CoursesController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Course"}}}},"404":{"description":"Course not found"}},"security":[{"bearer":[]}],"summary":"Get a single course with full details","tags":["courses"]},"patch":{"operationId":"CoursesController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCourseDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Course"}}}}},"security":[{"bearer":[]}],"summary":"[Admin] Update course details","tags":["courses"]},"delete":{"operationId":"CoursesController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"204":{"description":""}},"security":[{"bearer":[]}],"summary":"[Admin] Delete a course and all its assignments","tags":["courses"]}},"/assignments":{"get":{"operationId":"AssignmentsController_findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Assignment"}}}}}},"security":[{"bearer":[]}],"summary":"List all assignments with their course","tags":["assignments"]},"post":{"operationId":"AssignmentsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateAssignmentDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Assignment"}}}}},"security":[{"bearer":[]}],"summary":"[Admin] Create an assignment for a course","tags":["assignments"]}},"/assignments/{id}":{"get":{"operationId":"AssignmentsController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Assignment"}}}},"404":{"description":"Assignment not found"}},"security":[{"bearer":[]}],"summary":"Get a single assignment","tags":["assignments"]},"patch":{"operationId":"AssignmentsController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateAssignmentDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Assignment"}}}}},"security":[{"bearer":[]}],"summary":"[Admin] Update an assignment","tags":["assignments"]},"delete":{"operationId":"AssignmentsController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"204":{"description":""}},"security":[{"bearer":[]}],"summary":"[Admin] Delete an assignment","tags":["assignments"]}}},"info":{"title":"Student Learning System API","description":"REST API for managing students, courses, assignments, and enrollments.\n\n**Authentication:** Use `POST /auth/register` to create an account, then `POST /auth/login` to receive a JWT. Click **Authorize** and enter `Bearer <token>` to unlock admin endpoints.","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"RegisterDto":{"type":"object","properties":{"email":{"type":"string","format":"email","example":"admin@school.com"},"password":{"type":"string","minLength":6,"example":"password123"},"role":{"enum":["admin","student"],"type":"string","default":"student"}},"required":["email","password"]},"LoginDto":{"type":"object","properties":{"email":{"type":"string","format":"email","example":"admin@school.com"},"password":{"type":"string","example":"password123"}},"required":["email","password"]},"Assignment":{"type":"object","properties":{"id":{"type":"number"},"title":{"type":"string"},"dueDate":{"type":"string"},"course":{"$ref":"#/components/schemas/Course"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"}},"required":["id","title","dueDate","course","createdAt","updatedAt"]},"Course":{"type":"object","properties":{"id":{"type":"number"},"title":{"type":"string"},"code":{"type":"string"},"assignments":{"type":"array","items":{"$ref":"#/components/schemas/Assignment"}},"students":{"type":"array","items":{"$ref":"#/components/schemas/Student"}},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"}},"required":["id","title","code","assignments","students","createdAt","updatedAt"]},"Student":{"type":"object","properties":{"id":{"type":"number"},"name":{"type":"string"},"email":{"type":"string"},"profile":{"$ref":"#/components/schemas/Profile"},"courses":{"type":"array","items":{"$ref":"#/components/schemas/Course"}},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"},"deletedAt":{"format":"date-time","type":"string"}},"required":["id","name","email","profile","courses","createdAt","updatedAt","deletedAt"]},"Profile":{"type":"object","properties":{"id":{"type":"number"},"bio":{"type":"string"},"avatarUrl":{"type":"string"},"student":{"$ref":"#/components/schemas/Student"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"}},"required":["id","bio","avatarUrl","student","createdAt","updatedAt"]},"CreateStudentDto":{"type":"object","properties":{"name":{"type":"string","example":"Jane Doe","description":"Full name of the student"},"email":{"type":"string","format":"email","example":"jane@example.com","description":"Unique email address"}},"required":["name","email"]},"UpdateStudentDto":{"type":"object","properties":{"name":{"type":"string","example":"Jane Smith"},"email":{"type":"string","format":"email","example":"jane.smith@example.com"}}},"CreateProfileDto":{"type":"object","properties":{"studentId":{"type":"number","example":1,"description":"Student ID to attach the profile to"},"bio":{"type":"string","example":"Passionate about mathematics and science."},"avatarUrl":{"type":"string","example":"https://example.com/avatar.jpg"}},"required":["studentId"]},"UpdateProfileDto":{"type":"object","properties":{"bio":{"type":"string","example":"Loves art and history."},"avatarUrl":{"type":"string","example":"https://example.com/new-avatar.jpg"}}},"CreateCourseDto":{"type":"object","properties":{"title":{"type":"string","example":"Introduction to Computer Science","description":"Course title"},"code":{"type":"string","example":"CS101","description":"Unique course code"}},"required":["title","code"]},"UpdateCourseDto":{"type":"object","properties":{"title":{"type":"string","example":"Advanced Computer Science"},"code":{"type":"string","example":"CS201"}}},"CreateAssignmentDto":{"type":"object","properties":{"title":{"type":"string","example":"Homework 1","description":"Assignment title"},"dueDate":{"type":"string","example":"2025-12-31","description":"Due date in YYYY-MM-DD format"},"courseId":{"type":"number","example":1,"description":"ID of the course this assignment belongs to"}},"required":["title","courseId"]},"UpdateAssignmentDto":{"type":"object","properties":{"title":{"type":"string","example":"Homework 2"},"dueDate":{"type":"string","example":"2026-01-15"}}}}}}